Logs traceback errors on model (#18)
* add errors field

* log exception in view

* Resets migration dependency and file name

Changes migration file and dependency back to the initial migration file. The project moved forward with a backwards incompatible change and this PR was not updated to reflect that change. All migrations were squashed into the initial migration for simplicity.

* Fixes incorrect indentation to pass lint check

Through user error, the indentation of this test method was incorrectly added. Likely, due to merging conflicts from the master branch using the GIthub UI.

* fix(): remove wrong param in traceback.format_exc (#39)

* Updates and renames migration file

Updating the order number of the file and the migration dependency resolves the migration conflict in the PR.