Prevent log save failure to break API call
This is a follow-up of the discussion in PR #55.

If something is going wrong during log saving, we don't want
the API to fail and throw out an error to the end user.

To do this, silently swallow any exception throwed by save().

A test is provided to check that the response status_code of
the API is still 200 even in the event of a DB failure.